Math Duel: Play online
This game which includes the most effective mathematical methods of mental arithmetic.You can find math games, quiz and tests to improve your skills, measure your progress and stay on track with your goals. Take on the challenge and test your brain and mathematic abilities.
The game is also suitable for children,school kids, teaching in the form of the game allows you to develop an abstract and logical thinking - this is a very effective way to develop logical thinking in the form of a puzzle game.
We hope that you find it useful to improve your children’s knowledge and that it serve as a support for their studies.

Now , this game have multiplayer game mode. You can play with friend via internet .
Duel with friend or parent can play with their kid.

Test your mind, skill speed and exercise your brain once a day- play and have fun- Let's go